+++ base/exports.jl
@@ -494,7 +494,6 @@ export
     find,
     findfirst,
     findlast,
-    findin,
     findmax,
     findmin,
     findmin!,